test(voice-call): split events.test.ts under the max-lines cap
