+++ extensions/voice-call/src/manager.test-harness.ts
@@ -3,19 +3,25 @@ import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
 import type { OpenKeyedStoreOptions } from "openclaw/plugin-sdk/plugin-state-runtime";
-import { createPluginStateSyncKeyedStoreForTests } from "openclaw/plugin-sdk/plugin-state-test-runtime";
+import {
+  createPluginStateSyncKeyedStoreForTests,
+  resetPluginStateStoreForTests,
+} from "openclaw/plugin-sdk/plugin-state-test-runtime";
 import { VoiceCallConfigSchema } from "./config.js";
 import { CallManager } from "./manager.js";
+import type { CallManagerContext } from "./manager/context.js";
 import { persistCallRecord } from "./manager/store.js";
 import type { VoiceCallProvider } from "./providers/base.js";
 import { setVoiceCallStateRuntime, type VoiceCallStateRuntime } from "./runtime-state.js";
 import { CallRecordSchema } from "./types.js";
 import type {
+  CallRecord,
   GetCallStatusInput,
   GetCallStatusResult,
   HangupCallInput,
   InitiateCallInput,
   InitiateCallResult,
+  NormalizedEvent,
   PlayTtsInput,
   ProviderWebhookParseResult,
   StartListeningInput,
@@ -158,5 +164,157 @@ export function makePersistedCall(
     transcript: [],
     processedEventIds: [],
     ...overrides,
+  };
+}
+
+export const EVENT_MANAGER_REPLAY_KEY_LIMIT = 10_000;
+
+export function createEventManagerHarness() {
+  const contexts: CallManagerContext[] = [];
+
+  function installStateRuntime(shouldFail?: () => boolean): void {
+    setVoiceCallStateRuntime({
+      state: {
+        resolveStateDir: () => "",
+        openKeyedStore: (() => {
+          throw new Error("openKeyedStore is not used by voice-call event tests");
+        }) as never,
+        openSyncKeyedStore: (options: OpenKeyedStoreOptions) => {
+          if (shouldFail?.()) {
+            throw new Error("synthetic SQLite persistence failure");
+          }
+          return createPluginStateSyncKeyedStoreForTests("voice-call", options);
+        },
+        openChannelIngressQueue: (() => {
+          throw new Error("openChannelIngressQueue is not used by voice-call event tests");
+        }) as never,
+        openChannelIngressDrain: (() => {
+          throw new Error("openChannelIngressDrain is not used by voice-call event tests");
+        }) as never,
+      },
+    });
+  }
+
+  function setup(): void {
+    resetPluginStateStoreForTests();
+    installStateRuntime();
+  }
+
+  function cleanup(): void {
+    for (const ctx of contexts.splice(0)) {
+      for (const timer of ctx.maxDurationTimers.values()) {
+        clearTimeout(timer);
+      }
+      ctx.maxDurationTimers.clear();
+      for (const waiter of ctx.transcriptWaiters.values()) {
+        clearTimeout(waiter.timeout);
+      }
+      ctx.transcriptWaiters.clear();
+      fs.rmSync(ctx.storePath, { recursive: true, force: true });
+    }
+    resetPluginStateStoreForTests();
+  }
+
+  function createContext(overrides: Partial<CallManagerContext> = {}): CallManagerContext {
+    const storePath = fs.mkdtempSync(path.join(os.tmpdir(), "openclaw-voice-call-events-test-"));
+    const ctx: CallManagerContext = {
+      activeCalls: new Map(),
+      providerCallIdMap: new Map(),
+      processedEventIds: new Set(),
+      rejectedProviderCallIds: new Map(),
+      provider: null,
+      config: VoiceCallConfigSchema.parse({
+        enabled: true,
+        provider: "plivo",
+        fromNumber: "+15550000000",
+      }),
+      storePath,
+      webhookUrl: null,
+      activeTurnCalls: new Set(),
+      transcriptWaiters: new Map(),
+      maxDurationTimers: new Map(),
+      initialMessageInFlight: new Set(),
+      ...overrides,
+    };
+    contexts.push(ctx);
+    return ctx;
+  }
+
+  function createProvider(overrides: Partial<VoiceCallProvider> = {}): VoiceCallProvider {
+    return {
+      name: "plivo",
+      verifyWebhook: () => ({ ok: true }),
+      parseWebhookEvent: () => ({ events: [] }),
+      initiateCall: async () => ({ providerCallId: "provider-call-id", status: "initiated" }),
+      hangupCall: async () => {},
+      playTts: async () => {},
+      startListening: async () => {},
+      stopListening: async () => {},
+      getCallStatus: async () => ({ status: "in-progress", isTerminal: false }),
+      ...overrides,
+    };
+  }
+
+  function createInboundDisabledConfig() {
+    return VoiceCallConfigSchema.parse({
+      enabled: true,
+      provider: "plivo",
+      fromNumber: "+15550000000",
+      inboundPolicy: "disabled",
+    });
+  }
+
+  function createInboundInitiatedEvent(params: {
+    id: string;
+    providerCallId: string;
+    from: string;
+  }): NormalizedEvent {
+    return {
+      id: params.id,
+      type: "call.initiated",
+      callId: params.providerCallId,
+      providerCallId: params.providerCallId,
+      timestamp: Date.now(),
+      direction: "inbound",
+      from: params.from,
+      to: "+15550000000",
+    };
+  }
+
+  function createRejectingInboundContext(): {
+    ctx: CallManagerContext;
+    hangupCalls: HangupCallInput[];
+  } {
+    const hangupCalls: HangupCallInput[] = [];
+    const provider = createProvider({
+      hangupCall: async (input: HangupCallInput): Promise<void> => {
+        hangupCalls.push(input);
+      },
+    });
+    const ctx = createContext({
+      config: createInboundDisabledConfig(),
+      provider,
+    });
+    return { ctx, hangupCalls };
+  }
+
+  function requireFirstActiveCall(ctx: CallManagerContext): CallRecord {
+    const call = [...ctx.activeCalls.values()][0];
+    if (!call) {
+      throw new Error("expected one active call");
+    }
+    return call;
+  }
+
+  return {
+    cleanup,
+    createContext,
+    createInboundDisabledConfig,
+    createInboundInitiatedEvent,
+    createProvider,
+    createRejectingInboundContext,
+    installStateRuntime,
+    requireFirstActiveCall,
+    setup,
   };
 }
