+++ extensions/voice-call/src/manager/events.test.ts
@@ -1,23 +1,16 @@
 // Voice Call tests cover events plugin behavior.
-import fs from "node:fs";
-import os from "node:os";
-import path from "node:path";
-import type { OpenKeyedStoreOptions } from "openclaw/plugin-sdk/plugin-state-runtime";
-import {
-  createPluginStateSyncKeyedStoreForTests,
-  resetPluginStateStoreForTests,
-} from "openclaw/plugin-sdk/plugin-state-test-runtime";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { VoiceCallConfigSchema } from "../config.js";
+import {
+  createEventManagerHarness,
+  EVENT_MANAGER_REPLAY_KEY_LIMIT,
+} from "../manager.test-harness.js";
 import type { VoiceCallProvider } from "../providers/base.js";
-import { setVoiceCallStateRuntime } from "../runtime-state.js";
-import type { AnswerCallInput, CallRecord, HangupCallInput, NormalizedEvent } from "../types.js";
-import type { CallManagerContext } from "./context.js";
+import type { CallRecord, HangupCallInput, NormalizedEvent } from "../types.js";
 import { processEvent } from "./events.js";
 import { speakInitialMessage } from "./outbound.js";
 import { MAX_CALL_REPLAY_KEYS } from "./replay-keys.js";
 
-const MANAGER_REPLAY_KEY_LIMIT = 10_000;
 const logSpy = vi.hoisted(() => {
   const logEntries: string[] = [];
   return {
@@ -46,143 +39,26 @@ vi.mock("openclaw/plugin-sdk/runtime-env", async (importOriginal) => {
   };
 });
 
-const contexts: CallManagerContext[] = [];
-
-function installStateRuntime(shouldFail?: () => boolean): void {
-  setVoiceCallStateRuntime({
-    state: {
-      resolveStateDir: () => "",
-      openKeyedStore: (() => {
-        throw new Error("openKeyedStore is not used by voice-call event tests");
-      }) as never,
-      openSyncKeyedStore: (options: OpenKeyedStoreOptions) => {
-        if (shouldFail?.()) {
-          throw new Error("synthetic SQLite persistence failure");
-        }
-        return createPluginStateSyncKeyedStoreForTests("voice-call", options);
-      },
-      openChannelIngressQueue: (() => {
-        throw new Error("openChannelIngressQueue is not used by voice-call event tests");
-      }) as never,
-      openChannelIngressDrain: (() => {
-        throw new Error("openChannelIngressDrain is not used by voice-call event tests");
-      }) as never,
-    },
-  });
-}
+const {
+  cleanup,
+  createContext,
+  createInboundInitiatedEvent,
+  createProvider,
+  createRejectingInboundContext,
+  installStateRuntime,
+  requireFirstActiveCall,
+  setup,
+} = createEventManagerHarness();
 
 beforeEach(() => {
-  resetPluginStateStoreForTests();
-  installStateRuntime();
+  setup();
 });
 
-afterEach(async () => {
-  for (const ctx of contexts.splice(0)) {
-    for (const timer of ctx.maxDurationTimers.values()) {
-      clearTimeout(timer);
-    }
-    ctx.maxDurationTimers.clear();
-    for (const waiter of ctx.transcriptWaiters.values()) {
-      clearTimeout(waiter.timeout);
-    }
-    ctx.transcriptWaiters.clear();
-    fs.rmSync(ctx.storePath, { recursive: true, force: true });
-  }
-  resetPluginStateStoreForTests();
+afterEach(() => {
+  cleanup();
   vi.useRealTimers();
   vi.restoreAllMocks();
 });
-
-function createContext(overrides: Partial<CallManagerContext> = {}): CallManagerContext {
-  const storePath = fs.mkdtempSync(path.join(os.tmpdir(), "openclaw-voice-call-events-test-"));
-  const ctx: CallManagerContext = {
-    activeCalls: new Map(),
-    providerCallIdMap: new Map(),
-    processedEventIds: new Set(),
-    rejectedProviderCallIds: new Map(),
-    provider: null,
-    config: VoiceCallConfigSchema.parse({
-      enabled: true,
-      provider: "plivo",
-      fromNumber: "+15550000000",
-    }),
-    storePath,
-    webhookUrl: null,
-    activeTurnCalls: new Set(),
-    transcriptWaiters: new Map(),
-    maxDurationTimers: new Map(),
-    initialMessageInFlight: new Set(),
-    ...overrides,
-  };
-  contexts.push(ctx);
-  return ctx;
-}
-
-function createProvider(overrides: Partial<VoiceCallProvider> = {}): VoiceCallProvider {
-  return {
-    name: "plivo",
-    verifyWebhook: () => ({ ok: true }),
-    parseWebhookEvent: () => ({ events: [] }),
-    initiateCall: async () => ({ providerCallId: "provider-call-id", status: "initiated" }),
-    hangupCall: async () => {},
-    playTts: async () => {},
-    startListening: async () => {},
-    stopListening: async () => {},
-    getCallStatus: async () => ({ status: "in-progress", isTerminal: false }),
-    ...overrides,
-  };
-}
-
-function createInboundDisabledConfig() {
-  return VoiceCallConfigSchema.parse({
-    enabled: true,
-    provider: "plivo",
-    fromNumber: "+15550000000",
-    inboundPolicy: "disabled",
-  });
-}
-
-function createInboundInitiatedEvent(params: {
-  id: string;
-  providerCallId: string;
-  from: string;
-}): NormalizedEvent {
-  return {
-    id: params.id,
-    type: "call.initiated",
-    callId: params.providerCallId,
-    providerCallId: params.providerCallId,
-    timestamp: Date.now(),
-    direction: "inbound",
-    from: params.from,
-    to: "+15550000000",
-  };
-}
-
-function createRejectingInboundContext(): {
-  ctx: CallManagerContext;
-  hangupCalls: HangupCallInput[];
-} {
-  const hangupCalls: HangupCallInput[] = [];
-  const provider = createProvider({
-    hangupCall: async (input: HangupCallInput): Promise<void> => {
-      hangupCalls.push(input);
-    },
-  });
-  const ctx = createContext({
-    config: createInboundDisabledConfig(),
-    provider,
-  });
-  return { ctx, hangupCalls };
-}
-
-function requireFirstActiveCall(ctx: CallManagerContext) {
-  const call = [...ctx.activeCalls.values()][0];
-  if (!call) {
-    throw new Error("expected one active call");
-  }
-  return call;
-}
 
 describe("processEvent (functional)", () => {
   it.each(["speech", "answered", "terminal"] as const)(
@@ -248,137 +124,6 @@ describe("processEvent (functional)", () => {
       expect(ctx.activeCalls.has(call.callId)).toBe(kind !== "terminal");
     },
   );
-
-  it.each(["created", "rejected"] as const)(
-    "does not publish %s inbound calls before SQLite persistence succeeds",
-    (kind) => {
-      let failPersistence = true;
-      installStateRuntime(() => failPersistence);
-      const { ctx, hangupCalls } = createRejectingInboundContext();
-      ctx.config.inboundPolicy = kind === "created" ? "open" : "disabled";
-      const event = createInboundInitiatedEvent({
-        id: `event-durable-${kind}`,
-        providerCallId: `provider-durable-${kind}`,
-        from: "+15550000002",
-      });
-
-      expect(() => processEvent(ctx, event)).toThrow("synthetic SQLite persistence failure");
-      expect(ctx.activeCalls.size).toBe(0);
-      expect(ctx.providerCallIdMap.size).toBe(0);
-      expect(ctx.rejectedProviderCallIds.size).toBe(0);
-      expect(ctx.processedEventIds.size).toBe(0);
-      expect(hangupCalls).toHaveLength(0);
-
-      failPersistence = false;
-      expect(processEvent(ctx, event)).toEqual({ kind: "processed" });
-      expect(ctx.activeCalls.size).toBe(kind === "created" ? 1 : 0);
-      expect(hangupCalls).toHaveLength(kind === "rejected" ? 1 : 0);
-    },
-  );
-
-  it("calls provider hangup when rejecting inbound call", () => {
-    const { ctx, hangupCalls } = createRejectingInboundContext();
-    const event = createInboundInitiatedEvent({
-      id: "evt-1",
-      providerCallId: "prov-1",
-      from: "+15559999999",
-    });
-
-    processEvent(ctx, event);
-
-    expect(ctx.activeCalls.size).toBe(0);
-    expect(hangupCalls).toHaveLength(1);
-    expect(hangupCalls[0]).toEqual({
-      callId: "prov-1",
-      providerCallId: "prov-1",
-      reason: "hangup-bot",
-    });
-  });
-
-  it("does not call hangup when provider is null", () => {
-    const ctx = createContext({
-      config: createInboundDisabledConfig(),
-      provider: null,
-    });
-    const event = createInboundInitiatedEvent({
-      id: "evt-2",
-      providerCallId: "prov-2",
-      from: "+15551111111",
-    });
-
-    processEvent(ctx, event);
-
-    expect(ctx.activeCalls.size).toBe(0);
-  });
-
-  it("calls hangup only once for duplicate events for same rejected call", () => {
-    const { ctx, hangupCalls } = createRejectingInboundContext();
-    const event1 = createInboundInitiatedEvent({
-      id: "evt-init",
-      providerCallId: "prov-dup",
-      from: "+15552222222",
-    });
-    const event2: NormalizedEvent = {
-      id: "evt-ring",
-      type: "call.ringing",
-      callId: "prov-dup",
-      providerCallId: "prov-dup",
-      timestamp: Date.now(),
-      direction: "inbound",
-      from: "+15552222222",
-      to: "+15550000000",
-    };
-
-    processEvent(ctx, event1);
-    processEvent(ctx, event2);
-
-    expect(ctx.activeCalls.size).toBe(0);
-    expect(hangupCalls).toEqual([
-      {
-        callId: "prov-dup",
-        providerCallId: "prov-dup",
-        reason: "hangup-bot",
-      },
-    ]);
-  });
-
-  it("answers accepted inbound calls when the provider requires an answer command", () => {
-    const answerCalls: AnswerCallInput[] = [];
-    const provider = createProvider({
-      answerCall: async (input: AnswerCallInput): Promise<void> => {
-        answerCalls.push(input);
-      },
-    });
-    const ctx = createContext({
-      config: VoiceCallConfigSchema.parse({
-        enabled: true,
-        provider: "telnyx",
-        fromNumber: "+15550000000",
-        inboundPolicy: "open",
-        telnyx: {
-          apiKey: "KEY123",
-          connectionId: "CONN456",
-        },
-        skipSignatureVerification: true,
-      }),
-      provider,
-    });
-    const event = createInboundInitiatedEvent({
-      id: "evt-answer",
-      providerCallId: "call-control-1",
-      from: "+15552222222",
-    });
-
-    processEvent(ctx, event);
-
-    const call = requireFirstActiveCall(ctx);
-    expect(answerCalls).toEqual([
-      {
-        callId: call.callId,
-        providerCallId: "call-control-1",
-      },
-    ]);
-  });
 
   it("updates providerCallId map when provider ID changes", () => {
     const now = Date.now();
@@ -658,26 +403,6 @@ describe("processEvent (functional)", () => {
     vi.useRealTimers();
   });
 
-  it("removes active call even when hangup rejects", () => {
-    const provider = createProvider({
-      hangupCall: async (): Promise<void> => {
-        throw new Error("provider down");
-      },
-    });
-    const ctx = createContext({
-      config: createInboundDisabledConfig(),
-      provider,
-    });
-    const event = createInboundInitiatedEvent({
-      id: "evt-fail",
-      providerCallId: "prov-fail",
-      from: "+15553333333",
-    });
-
-    processEvent(ctx, event);
-    expect(ctx.activeCalls.size).toBe(0);
-  });
-
   it("auto-registers externally-initiated outbound-api calls with correct direction", () => {
     const ctx = createContext();
     const event: NormalizedEvent = {
@@ -723,110 +448,6 @@ describe("processEvent (functional)", () => {
     expect(hangupCalls).toHaveLength(0);
     const call = requireFirstActiveCall(ctx);
     expect(call.direction).toBe("outbound");
-  });
-
-  it("preserves inbound direction for auto-registered inbound calls", () => {
-    const ctx = createContext({
-      config: VoiceCallConfigSchema.parse({
-        enabled: true,
-        provider: "plivo",
-        fromNumber: "+15550000000",
-        inboundPolicy: "open",
-      }),
-    });
-    const event: NormalizedEvent = {
-      id: "evt-inbound-dir",
-      type: "call.initiated",
-      callId: "CA-inbound-789",
-      providerCallId: "CA-inbound-789",
-      timestamp: Date.now(),
-      direction: "inbound",
-      from: "+15554444444",
-      to: "+15550000000",
-    };
-
-    processEvent(ctx, event);
-
-    expect(ctx.activeCalls.size).toBe(1);
-    const call = requireFirstActiveCall(ctx);
-    expect(call.direction).toBe("inbound");
-  });
-
-  it.each([
-    {
-      sessionScope: "per-call",
-      coreSession: undefined,
-      expectedSessionKey: (call: CallRecord) => `agent:main:voice:call:${call.callId}`,
-    },
-    {
-      sessionScope: "main",
-      coreSession: { mainKey: "work" },
-      expectedSessionKey: () => "agent:main:work",
-    },
-  ])(
-    "assigns $sessionScope session keys to inbound calls",
-    ({ sessionScope, coreSession, expectedSessionKey }) => {
-      const ctx = createContext({
-        config: VoiceCallConfigSchema.parse({
-          enabled: true,
-          provider: "plivo",
-          fromNumber: "+15550000000",
-          inboundPolicy: "open",
-          sessionScope,
-        }),
-        coreSession,
-      });
-      const event: NormalizedEvent = {
-        id: "evt-inbound-session-scope",
-        type: "call.initiated",
-        callId: "CA-inbound-session-scope",
-        providerCallId: "CA-inbound-session-scope",
-        timestamp: Date.now(),
-        direction: "inbound",
-        from: "+15554444444",
-        to: "+15550000000",
-      };
-
-      processEvent(ctx, event);
-
-      const call = requireFirstActiveCall(ctx);
-      expect(call.sessionKey).toBe(expectedSessionKey(call));
-    },
-  );
-
-  it("applies per-number inbound greeting and stores the matched route key", () => {
-    const ctx = createContext({
-      config: VoiceCallConfigSchema.parse({
-        enabled: true,
-        provider: "plivo",
-        fromNumber: "+15550000000",
-        inboundPolicy: "open",
-        inboundGreeting: "Hello from global.",
-        numbers: {
-          "+15550002222": {
-            agentId: "cards",
-            inboundGreeting: "Silver Fox Cards, how can I help?",
-          },
-        },
-      }),
-    });
-    const event: NormalizedEvent = {
-      id: "evt-inbound-number-route",
-      type: "call.initiated",
-      callId: "CA-inbound-number-route",
-      providerCallId: "CA-inbound-number-route",
-      timestamp: Date.now(),
-      direction: "inbound",
-      from: "+15554444444",
-      to: "+1 (555) 000-2222",
-    };
-
-    processEvent(ctx, event);
-
-    const call = requireFirstActiveCall(ctx);
-    expect(call.metadata?.initialMessage).toBe("Silver Fox Cards, how can I help?");
-    expect(call.metadata?.numberRouteKey).toBe("+15550002222");
-    expect(call.agentId).toBe("cards");
   });
 
   it("deduplicates by dedupeKey even when event IDs differ", () => {
@@ -886,7 +507,7 @@ describe("processEvent (functional)", () => {
   it("bounds committed replay keys in both manager and persisted call owners", () => {
     const now = Date.now();
     const managerKeys = Array.from(
-      { length: MANAGER_REPLAY_KEY_LIMIT },
+      { length: EVENT_MANAGER_REPLAY_KEY_LIMIT },
       (_, index) => `manager-${index}`,
     );
     const callKeys = Array.from({ length: MAX_CALL_REPLAY_KEYS }, (_, index) => `call-${index}`);
@@ -917,7 +538,7 @@ describe("processEvent (functional)", () => {
 
     const call = ctx.activeCalls.get("call-bounded");
     expect(result).toEqual({ kind: "processed" });
-    expect(ctx.processedEventIds.size).toBe(MANAGER_REPLAY_KEY_LIMIT);
+    expect(ctx.processedEventIds.size).toBe(EVENT_MANAGER_REPLAY_KEY_LIMIT);
     expect(ctx.processedEventIds.has("manager-0")).toBe(false);
     expect(ctx.processedEventIds.has("evt-bounded-new")).toBe(true);
     expect(call?.processedEventIds).toHaveLength(MAX_CALL_REPLAY_KEYS);
@@ -933,39 +554,6 @@ describe("processEvent (functional)", () => {
         digits: "1",
       }),
     ).toEqual({ kind: "ignored" });
-  });
-
-  it("bounds rejected provider calls while retaining hangup-once behavior", () => {
-    const rejectedProviderCallIds = new Map<string, symbol>(
-      Array.from(
-        { length: MANAGER_REPLAY_KEY_LIMIT },
-        (_, index) => [`provider-${index}`, Symbol(`provider-${index}`)] as const,
-      ),
-    );
-    const { ctx, hangupCalls } = createRejectingInboundContext();
-    ctx.rejectedProviderCallIds = rejectedProviderCallIds;
-
-    processEvent(
-      ctx,
-      createInboundInitiatedEvent({
-        id: "evt-rejected-new",
-        providerCallId: "provider-new",
-        from: "+15552222222",
-      }),
-    );
-    processEvent(
-      ctx,
-      createInboundInitiatedEvent({
-        id: "evt-rejected-new-replay",
-        providerCallId: "provider-new",
-        from: "+15552222222",
-      }),
-    );
-
-    expect(ctx.rejectedProviderCallIds.size).toBe(MANAGER_REPLAY_KEY_LIMIT);
-    expect(ctx.rejectedProviderCallIds.has("provider-0")).toBe(false);
-    expect(ctx.rejectedProviderCallIds.has("provider-new")).toBe(true);
-    expect(hangupCalls).toHaveLength(1);
   });
 
   it("keeps retryable call.error events replayable", () => {
